+++ tests/testthat/test-parallel-chat.R
@@ -79,31 +79,33 @@ test_that("can extract data in parallel", {
   chat <- chat_openai_test()
   data <- parallel_chat_structured(
     chat,
-    list(
-      "John, age 15, won first prize",
-      "Jane, age 16, won second prize"
-    ),
+    list("John, age 15", "Jane, age 16"),
     type = person
   )
   expect_equal(data, data.frame(name = c("John", "Jane"), age = c(15, 16)))
 })
 
-test_that("can get tokens and/or cost", {
-  vcr::local_cassette("parallel-data-cost")
-
-  # These are pretty weak, but it's hard to know how to do better.
+test_that("can get tokens", {
+  vcr::local_cassette("parallel-data")
   person <- type_object(name = type_string(), age = type_integer())
-
   chat <- chat_openai_test()
+
   data <- parallel_chat_structured(
     chat,
     list("John, age 15", "Jane, age 16"),
     type = person,
     include_tokens = TRUE
   )
+  # These are pretty weak, but it's hard to know how to do better.
   expect_contains(names(data), c("input_tokens", "output_tokens"))
   expect_equal(data$input_tokens > 0, c(TRUE, TRUE))
   expect_equal(data$output_tokens > 0, c(TRUE, TRUE))
+})
+
+test_that("can get cost", {
+  vcr::local_cassette("parallel-data")
+  person <- type_object(name = type_string(), age = type_integer())
+  chat <- chat_openai_test()
 
   data <- parallel_chat_structured(
     chat,
@@ -113,6 +115,12 @@ test_that("can get tokens and/or cost", {
   )
   expect_contains(names(data), "cost")
   expect_equal(data$cost > 0, c(TRUE, TRUE))
+})
+
+test_that("can get tokens & cost", {
+  vcr::local_cassette("parallel-data")
+  person <- type_object(name = type_string(), age = type_integer())
+  chat <- chat_openai_test()
 
   data <- parallel_chat_structured(
     chat,