+++ NEWS.md
@@ -1,5 +1,6 @@
 # ellmer (development version)
 
+* `Chat$chat_structured()` and friends no longer unnecessarily wrap `type_object()` for `chat_openai()` (#671)
 * `Chat$chat_structured()` now suppresses tool use. If you need to use tools and structured data together, first `$chat()` to use any tools needed, and then `$chat_structured()` to extract the data you need.
 * Now use prices data aggregated by LiteLLM. This considerably expands the number of providers and models that include cost information (#659).
 * `chat_aws_bedrock()` now allows you to set the `base_url()` (#441).